snapshot+net: make devices resumable and reinit TAP

+++ src/networking.h
@@ -73,7 +73,7 @@ int32_t net_tcp_recv(net_sock_t* sock, void* buffer, size_t size);
 
 net_sock_t* net_udp_bind(const net_addr_t* addr);
 
-size_t  net_udp_send(net_sock_t* sock, const void* buffer, size_t size, const net_addr_t* addr);
+int32_t net_udp_send(net_sock_t* sock, const void* buffer, size_t size, const net_addr_t* addr);
 int32_t net_udp_recv(net_sock_t* sock, void* buffer, size_t size, net_addr_t* addr);
 
 // ICMP Datagram Sockets (Optional)
@@ -81,7 +81,7 @@ int32_t net_udp_recv(net_sock_t* sock, void* buffer, size_t size, net_addr_t* ad
 net_sock_t* net_icmp_bind(const net_addr_t* addr);
 uint16_t    net_icmp_id(net_sock_t* sock);
 
-size_t  net_icmp_send(net_sock_t* sock, const void* buffer, size_t size, const net_addr_t* addr);
+int32_t net_icmp_send(net_sock_t* sock, const void* buffer, size_t size, const net_addr_t* addr);
 int32_t net_icmp_recv(net_sock_t* sock, void* buffer, size_t size, net_addr_t* addr);
 
 // Generic socket operations
